fix: coverage and minor refactoring
