+++ src/components/games/blackjack/useBlackjackAI.ts
@@ -6,20 +6,17 @@ import blackjackState, {
 } from "../../../jotai/blackjack-state";
 import { DBHand, DBPlayer } from "../../../jotai/player-atom";
 import { TurnState } from "../../../jotai/turn-atom";
-import { DBCard, rankSort } from "../../../jotai/deck-state";
+import { rankSort } from "../../../jotai/deck-state";
 import useDeck from "./api/useDeck";
 import asyncForEach from "./api/asyncForEach";
-import { getGameFunctions } from "./blackjackHelpers";
+import {
+  banking,
+  DEALER,
+  getGameFunctions,
+  weighHand,
+} from "./blackjackHelpers";
 
-// Dealer constant
-const DEALER = 0;
 const D_H_TURN = 0;
-
-interface PlayerStats {
-  house: number;
-  payout: number;
-  status: string;
-}
 
 interface FullBlackJackState {
   turn: TurnState;
@@ -31,106 +28,6 @@ interface PlayBotState {
   hands: DBHand[];
   doubled?: boolean;
 }
-
-/* ------------------------------     Pure Functions     ------------------------------ */
-
-/** calculate the weight of a hand */
-export function weighHand(hand: DBCard[] = []): {
-  weight: number;
-  soft: boolean;
-} {
-  // set return values
-  let weight = 0;
-  let soft = false;
-
-  // find the weight of the hand
-  hand.forEach((card) => {
-    const cardWeight = card.weight || 0;
-    if (cardWeight === 14) {
-      // A
-      if (weight <= 10) {
-        weight += 11;
-        soft = true;
-      } else {
-        weight += 1;
-      }
-    } else if (cardWeight > 10) {
-      // J - K
-      weight += 10;
-    } else {
-      // 2 - 10
-      weight += cardWeight;
-    }
-    // reduce by 10 if bust and soft
-    if (weight > 21 && soft) {
-      weight -= 10;
-      soft = false;
-    }
-  });
-
-  // return object w/ useful information
-  return { weight, soft };
-}
-
-/** finish the game and check for a winner */
-const banking = (players: DBPlayer[]): DBPlayer[] => {
-  // state variables
-  const dealer = players.filter((p) => p.id === DEALER)[0];
-  const dWeight = dealer.hands[0].weight || 0;
-  const dLength = dealer.hands[0].cards.length;
-  // track and find the winners
-  const playerStats: PlayerStats = { house: 0, payout: 0, status: "" };
-  // helper functions
-  const win = (ps: PlayerStats, bet: number, mul = 1) => {
-    ps.house -= Math.floor(mul * bet);
-    ps.payout = Math.floor(mul * bet);
-    ps.status = "win";
-  };
-  const loss = (ps: PlayerStats, bet: number) => {
-    ps.house += bet;
-    ps.payout = -bet;
-    ps.status = "lose";
-  };
-  return players.map((player) => {
-    const { id, bet } = player;
-    if (id === DEALER) {
-      if (playerStats.house > 0) {
-        playerStats.status = "win";
-      } else if (playerStats.house < 0) {
-        playerStats.status = "lose";
-      } else {
-        playerStats.status = "push";
-      }
-      return {
-        ...player,
-        status: playerStats.status,
-        money: player.money + playerStats.house,
-      };
-    }
-    player.hands.forEach((hand) => {
-      const { weight = 0, cards } = hand;
-      if (dWeight === 21 && dLength === 2) {
-        // dealer BlackJack
-        loss(playerStats, bet);
-      } else if (weight === 21 && cards.length === 2) {
-        // player BlackJack
-        win(playerStats, bet, 6 / 5);
-      } else if (weight <= 21 && (weight > dWeight || dWeight > 21)) {
-        win(playerStats, bet);
-      } else if (weight <= 21 && weight === dWeight) {
-        playerStats.payout = 0;
-        playerStats.status = "push";
-      } else {
-        loss(playerStats, bet);
-      }
-    });
-    return {
-      ...player,
-      status: playerStats.status,
-      money: player.money + playerStats.payout,
-    };
-  });
-};
 
 /* ------------------------------     Custom Hook     ------------------------------ */
 const useBlackjackAI = () => {
