+++ src/components/games/botc/botcHelpers.ts
