+++ src/components/resume/work/Job.tsx
@@ -3,7 +3,7 @@ import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
 import ExpandableCard from "../../common/expandable-card";
 import { Job as JobType } from "../../../constants/work";
-import { getCSV, showRange } from "./job-helpers";
+import { getCSV, showRange } from "./jobHelpers";
 
 interface JobProps {
   job: JobType;
