+++ src/components/resume/work/jobHelpers.tsx
